requirements.txt
joblib==1.4.2
spacy==3.7.5
streamlit==1.35.0
pandas==2.2.2
altair==5.3.0
streamlit-navigation-bar==3.3.0
hydralit-components==1.0.10
python-dotenv==1.0.1
scikit-learn==1.5.0
imbalanced-learn==0.12.3
pygithub==2.3.0
groq==0.9.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz#egg=en_core_web_sm
